+++ amd_openvx_extensions/amd_rpp/include/vx_ext_rpp.h
@@ -840,13 +840,14 @@ extern "C"
 	 * \param [out] pDst The output tensor in <tt>\ref VX_TYPE_UINT8</tt> or <tt>\ref VX_TYPE_FLOAT32</tt> or <tt>\ref VX_TYPE_FLOAT16</tt> or <tt>\ref VX_TYPE_INT8</tt> format data.
 	 * \param [in] pStdDev The input array in <tt>\ref VX_TYPE_FLOAT32</tt> format containing the standard deviation values for Gaussian filter kernel generation.
 	 * \param [in] kernelSize The input scalar in <tt>\ref VX_TYPE_UINT32</tt> format containing the kernel size for the Gaussian filter.
+	 * \param [in] borderType The input scalar in <tt>\ref VX_TYPE_INT32</tt> format containing the border type for handling edge pixels.
 	 * \param [in] inputLayout The input layout in <tt>\ref VX_TYPE_INT32</tt> denotes the layout of input tensor.
 	 * \param [in] outputLayout The output layout in <tt>\ref VX_TYPE_INT32</tt> denotes the layout of output tensor.
 	 * \param [in] roiType The type of roi <tt>\ref VX_TYPE_INT32</tt> denotes whether source roi is of XYWH/LTRB type.
 	 * \return A node reference <tt>\ref vx_node</tt>. Any possible errors preventing a successful creation should be checked using <tt>\ref vxGetStatus</tt>.
 	 */
 	SHARED_PUBLIC vx_node VX_API_CALL vxExtRppGaussianFilter(vx_graph graph, vx_tensor pSrc, vx_tensor pSrcRoi, vx_tensor pDst,
-	                                                         vx_array pStdDev, vx_scalar kernelSize,
+	                                                         vx_array pStdDev, vx_scalar kernelSize, vx_scalar borderType,
 	                                                         vx_scalar inputLayout, vx_scalar outputLayout, vx_scalar roiType);
 #ifdef __cplusplus
 }